+++ tests/archives/nextdom_tests/bootstrap.php
@@ -1,0 +1,3 @@
+<?php
+echo "Test de NextDom\n";
+?>
