+++ tests/archives/old_tests/bootstrap.php
@@ -1,0 +1,4 @@
+<?php
+require_once __DIR__ . "/../core/php/core.inc.php";
+echo "\nCommence le test NextDom\n";
+?>
