--- tests/nextdom_tests/ClientTest.php
@@ -1,50 +1,0 @@
-<?php
-/* This file is part of NextDom.
- *
- * NextDom is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * NextDom is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with NextDom. If not, see <http://www.gnu.org/licenses/>.
- */
-
-use NextDom\Helpers\Client;
-
-class ClientTest extends PHPUnit_Framework_TestCase
-{
-    public function testIsMobileWithoutUserAgent()
-    {
-        $this->assertFalse(Client::isMobile());
-    }
-
-    public function testIsMobileWithEmptyUserAgent()
-    {
-        $_SERVER['HTTP_USER_AGENT'] = '';
-        $this->assertFalse(Client::isMobile());
-    }
-
-    public function testIsMobileDesktop()
-    {
-        $_SERVER['HTTP_USER_AGENT'] = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_5) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/66.0.3359.181 Safari/537.36';
-        $this->assertFalse(Client::isMobile());
-    }
-
-    public function testIsMobileAndroid()
-    {
-        $_SERVER['HTTP_USER_AGENT'] = 'Mozilla/5.0 (Linux; U; Android 4.4.2; en-us; SCH-I535 Build/KOT49H) AppleWebKit/534.30 (KHTML, like Gecko) Version/4.0 Mobile Safari/534.30';
-        $this->assertTrue(Client::isMobile());
-    }
-
-    public function testIsMobileiPhone()
-    {
-        $_SERVER['HTTP_USER_AGENT'] = 'Mozilla/5.0 (iPhone; CPU iPhone OS 10_3_1 like Mac OS X) AppleWebKit/603.1.30 (KHTML, like Gecko) Version/10.0 Mobile/14E304 Safari/602.1';
-        $this->assertTrue(Client::isMobile());
-    }
-}
