--- tests/nextdom_tests/bootstrap.php
@@ -1,3 +1,0 @@
-<?php
-echo "Test de NextDom\n";
-?>
