--- tests/old_tests/bootstrap.php
@@ -1,4 +1,0 @@
-<?php
-require_once __DIR__ . "/../core/php/core.inc.php";
-echo "\nCommence le test NextDom\n";
-?>
