+++ tests/scripts/remove_docker.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+
+docker kill nextdom-test > /dev/null 2>&1
+docker rm nextdom-test > /dev/null 2>&1
+docker rmi nextdom-test-snap
